Add about section
With these changes we added an inital layout for the about section.
We created about.html to define the base layout including the background
image.
We also added the route to the about section in main.py.
We added a new image compass-sea.jpg to be used as the background image.
We updated the header to include a link to the about section.