Add a section to see all the holiday journals
Adds a section to see all the holiday journals. The section includes
a preview of each journal with the title, trip description, and date of
the trip.
Removes the unecessary logout link from the header.
Adds the holiday_journal link to the header
Add dummy list of journal to display in main.py.
Adds the routing to main.py and the html template to display.