Add database
With these changes we added a database to the app using SQLAlchemy. This
database will allow to store the user, holiday journal and data blog posts
tables. We created a package for the project where we added `database.py`
where the code relating to the database is stored.
We also added SQLAlchemy and flask_sqlalchemy to the list of dependencies.